Revert "Convert Main.java to JUnit test"
Remembered another time handing in maybe this exact same task. I got the complaint that they had to install IntelliJ to verify my work. They would rather just javac the code.

Besides: refactoring is to change the structure without changing external behaviour. And the test Main class can be seen as *the* external-facing behaviour of this project, so I should never have removed it. (But I will not use it as my automated tests runner. It's too clunky as a feedback loop.)

I'm sorry I had to change the instructions for running it, though. With the code in a package, it will no longer work to compile the code unless you first cd into the src/ directory. I suppose that counts as changing external behaviour, but I think it's worth it. Java code should have packages.

All the public accessors I removed now had to be replaced. That's ok, it's no big job with a decent IDE. I like “fixing” what the IDE tells me is “wrong.” Even if it's not even flagged with a warning. In this case it's about visibility/access. I don't regret reducing the scope of the classes (even if I had to revert it now). Setting the scope is a powerful tool for clarity, simplicity and reducing coupling. You cannot have any coupling to what is not visible. You can usually ignore what you cannot access when trying to understand your little part of the code (if you can assume good behaviour in the things you do access).