Rename the test class and move it to tests/
I had missed the fact that it was in the wrong folder. IntelliJ has a bad default when moving files into packages. Because I had first moved the src/ classes into the package, it chose that as the destination when I did the same for the tests. I didn't remember to check for that. Maybe, if I had done it in the opposite order, it would have been more obviously wrong?