Move implementation details to a new package
I like to separate concerns in this way. I also think it matches CCP and CSP to have a package (or better maybe: a separate JAR) for the details. This makes it easier e.g. to switch to a database later or to some other implementation whatever it might be.

Note: I think the Movie class is a detail, because the type of book is a String here and that is maybe useful in a database, but not in business logic. I will probably remove this class later, and I might add a new Movie class that fits the domain model (when the code is in a better state).