Remove getters from the Customer data structure
I dislike direct getters and setters. They do not truly encapsulate their respective fields. Setters are obviously worse as the represent mutation,  but getters are pretty bad too as they do not do their job of information hiding. That's what makes them pointless. It may be more intent-revealing to access the fields directly.

There may be two reason for keeping them around. One is that you may need them to implement an interface or an abstract class, and (in Java) that is only possible for methods.

The other is that it will couple you to the details, and it may cause a tsunami of change. I argue that when you have direct accessors, you already have data structures and barely any encapsulation, so will probably encounter the tsunami anyway.